INTERNAL: Move the box argument processor to its rightful place
Converted the Pallet namedtuple into a dataclass that will be expanded
upon in a later changeset. Nothing currently uses tuple unpacking with
the namedtuple returned from get_pallets, so this shouldn't change
behavior.

Changed the box argument processor to use the same Pallet dataclass as
the pallet argument processor when returning pallet information.
