+++ common/src/stack/command/stack/commands/__init__.py
@@ -168,60 +168,6 @@ class ApplianceArgumentProcessor:
 		return appliances
 
 
-class BoxArgumentProcessor:
-	"""
-	An Interface class to add the ability to process box arguments.
-	"""
-
-	def getBoxNames(self, args=None):
-		"""
-		Returns a list of box names from the database. For each arg in
-		the ARGS list find all the box names that match the arg (assume
-		SQL regexp). If an arg does not match anything in the database we
-		raise an exception. If the ARGS list is empty return all box names.
-		"""
-
-		boxes = []
-		if not args:
-			args = ['%']		      # find all boxes
-
-		for arg in args:
-			names = flatten(self.db.select(
-				'name from boxes where name like %s', (arg,)
-			))
-
-			if not names and arg != '%':
-				raise ArgNotFound(self, arg, 'box')
-
-			boxes.extend(names)
-
-		return boxes
-
-	def getBoxPallets(self, box='default'):
-		"""
-		Returns a list of pallets for a box
-		"""
-
-		# Make sure 'box' exists
-		self.getBoxNames([box])
-
-		pallets = []
-
-		Pallet = namedtuple(
-			"Pallet", ['id', 'name', 'version', 'rel', 'arch', 'os', 'url']
-		)
-
-		rows = self.db.select("""
-			r.id, r.name, r.version, r.rel, r.arch, o.name, r.url
-			from rolls r, boxes b, stacks s, oses o
-			where b.name=%s and b.id=s.box and s.roll=r.id and b.os=o.id
-		""", (box,))
-
-		pallets.extend([Pallet(*row) for row in rows])
-
-		return pallets
-
-
 class NetworkArgumentProcessor:
 	"""
 	An Interface class to add the ability to process network (subnet) argument.
