+++ common/src/stack/command/stack/commands/report/host/repo/imp_sles.py
@@ -18,7 +18,7 @@ class Implementation(stack.commands.Implementation):
 		filename = '/etc/zypp/repos.d/stacki.repo'
 		repo.append('<stack:file stack:name="%s">' % filename)
 
-		for pallet in self.owner.getBoxPallets(box):
+		for pallet in self.owner.get_box_pallets(box):
 
 			repo.append('[%s-%s-%s]' % (pallet.name, pallet.version, pallet.rel))
 			repo.append('name=%s %s %s' % (pallet.name, pallet.version, pallet.rel))
